# libs/contact/config/src/contact.api-spec.json
{
  "openapi": "3.0.0",
...
  "info": {
    "title": "Contact APIs",
    "version": "1",
    "description": "A collection of APIs used to get and set contact related data ",
    "contact": {
      "name": "Patrick Sullivan",
      "email": "Patrick.Joseph.Sullivan@protonmail.com",
      "url": "https://sullivanpj.github.io/open-system/services/contact"
    },
    "termsOfService": "https://sullivanpj.github.io/open-system/services/contact",
    "license": {
      "name": "BSD 2-Clause License",
      "url": "https://opensource.org/licenses/BSD-2-Clause"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development"
    }
  ],
  "paths": {
...
    "/contacts/{id}": {
      "parameters": [
        {
          "$ref": "../../../shared/config/src/shared.api-spec.json#/components/parameters/id"
        }
      ],
      "get": {
        "summary": "Get Contact By Id",
        "operationId": "getContactById",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactRecord"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/ProblemDetailsResponse"
                },
                "examples": {
                  "Error Response": {
                    "value": {
                      "type": "https://example.com/probs/out-of-credit",
                      "title": "You do not have enough credit.",
                      "detail": "Your current balance is 30, but that costs 50.",
                      "instance": "/account/12345/msgs/abc"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/ProblemDetailsResponse"
                },
                "examples": {
                  "Error Response": {
                    "value": {
                      "type": "https://example.com/probs/out-of-credit",
                      "title": "You do not have enough credit.",
                      "detail": "Your current balance is 30, but that costs 50.",
                      "instance": "/account/12345/msgs/abc"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/ProblemDetailsResponse"
                },
                "examples": {
                  "Error Response": {
                    "value": {
                      "type": "https://example.com/probs/out-of-credit",
                      "title": "You do not have enough credit.",
                      "detail": "Your current balance is 30, but that costs 50.",
                      "instance": "/account/12345/msgs/abc"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/ProblemDetailsResponse"
                },
                "examples": {
                  "Error Response": {
                    "value": {
                      "type": "https://example.com/probs/out-of-credit",
                      "title": "You do not have enough credit.",
                      "detail": "Your current balance is 30, but that costs 50.",
                      "instance": "/account/12345/msgs/abc"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [],
        "description": "An end point that returns data for a specific contact",
        "tags": [
          "Contact"
        ]
      },
      "put": {
        "summary": "Update Contact",
        "operationId": "updateContact",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/CommandSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/ProblemDetailsResponse"
                },
                "examples": {
                  "Error Response": {
                    "value": {
                      "type": "https://example.com/probs/out-of-credit",
                      "title": "You do not have enough credit.",
                      "detail": "Your current balance is 30, but that costs 50.",
                      "instance": "/account/12345/msgs/abc"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/ProblemDetailsResponse"
                },
                "examples": {
                  "Error Response": {
                    "value": {
                      "type": "https://example.com/probs/out-of-credit",
                      "title": "You do not have enough credit.",
                      "detail": "Your current balance is 30, but that costs 50.",
                      "instance": "/account/12345/msgs/abc"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/ProblemDetailsResponse"
                },
                "examples": {
                  "Error Response": {
                    "value": {
                      "type": "https://example.com/probs/out-of-credit",
                      "title": "You do not have enough credit.",
                      "detail": "Your current balance is 30, but that costs 50.",
                      "instance": "/account/12345/msgs/abc"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/ProblemDetailsResponse"
                },
                "examples": {
                  "Error Response": {
                    "value": {
                      "type": "https://example.com/probs/out-of-credit",
                      "title": "You do not have enough credit.",
                      "detail": "Your current balance is 30, but that costs 50.",
                      "instance": "/account/12345/msgs/abc"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Contact"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactHeaderRecord"
              }
            }
          },
          "description": ""
        },
        "description": "An end point that updates an existing contact",
        "parameters": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContactRecord": {
        "allOf": [
          {
            "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/RecordBase"
          },
          {
            "$ref": "#/components/schemas/ContactHeaderRecord"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactDetailRecord"
            }
          }
        ],
        "title": "ContactRecord",
        "description": "A full contact model including all of it's previously submitted details",
        "x-stoplight": {
          "id": "9ee78c1890acc"
        }
      },
      "ContactHeaderRecord": {
        "allOf": [
          {
            "properties": {
              "firstName": {
                "type": "string",
                "maxLength": 50
              },
              "lastName": {
                "type": "string",
                "maxLength": 50
              },
              "phoneNumber": {
                "type": "string",
                "maxLength": 15,
                "pattern": "^[\\+]?[(]?[0-9]{3}[)]?[-\\s\\.]?[0-9]{3}[-\\s\\.]?[0-9]{4,6}$"
              },
              "email": {
                "type": "string",
                "description": "The email address of the contact",
                "format": "email"
              },
              "isSubscribed": {
                "type": "boolean",
                "default": true
              }
            },
            "required": [
              "email",
              "reason",
              "isSubscribed"
            ]
          },
          {
            "$ref": "../../../shared/config/src/shared.api-spec.json#/components/schemas/Address"
          }
        ],
        "type": "object",
        "title": "ContactHeader",
        "description": "Contact data used to identify an individual and attach details too.",
        "x-stoplight": {
          "id": "38655c7314ec4"
        }
      },
      "ContactDetailRecord": {
        "type": "object",
        "title": "ContactDetail",
        "description": "Base data used to capture a contact request ",
        "properties": {
          "companyName": {
            "type": "string",
            "maxLength": 50
          },
          "title": {
            "type": "string",
            "maxLength": 50
          },
          "reason": {
            "type": "string",
            "example": "business",
            "description": "The type of contact request",
            "default": "business",
            "enum": [
              "business",
              "question",
              "other",
              "project",
              "interest",
              "subscription"
            ]
          },
          "details": {
            "type": "string",
            "maxLength": 2000
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "reason"
        ],
        "x-stoplight": {
          "id": "63946c2ead1c5"
        }
      }
    },
...
    "securitySchemes": {}
  },
...
}